refactor: Add a new component for seeding posts functionality
